feat : add auth to chat / move socket directory
