+++ back/server.js
@@ -1,7 +1,7 @@
 const express = require('express')
 const http = require('http')
 const SocketIO = require('socket.io')
-const socket = require('./public/js/socket.js')
+const socket = require('./socket.js')
 const cors = require('cors')
 const cookieParser = require('cookie-parser')
 const router = require('./routes/index.js')
