+++ front/public/js/chat/popup.js
@@ -8,14 +8,5 @@ function openWindowPop(url, name){
     window.open(url, name, options)
 }
 
-const roomMaker = document.querySelector('#roomMaker')
-const roomSubmit = document.querySelector('#roomSubmit')
-roomMaker.addEventListener('click', ()=>{
-    if (roomSubmit.classList.value == 'on') {
-        roomSubmit.classList.remove('on')
-    } else {
-        roomSubmit.setAttribute('class', 'on')
-    }
-})
