+++ front/routes/chat/chatController.js
@@ -1,6 +1,7 @@
 const axios = require('axios')
 
 const getChatList = async (req, res) => {
+    const { userid } = req.userInfo
     try {
         const response = await axios.get('http://localhost:4001/api/chat/list', {
             withCredentials: true
@@ -8,7 +9,8 @@ const getChatList = async (req, res) => {
         const result = response.data.rows
     
         res.render('./chat/chat.html',{
-            result
+            result,
+            userid
         })
     } catch (err) {
         console.log(err)
@@ -17,9 +19,11 @@ const getChatList = async (req, res) => {
 }
 
 const getChatRoom = (req, res) => {
+    const { nickname } = req.userInfo
     const roomName = req.query.name
     res.render('./chat/chatRoom.html', {
-        roomName
+        roomName,
+        nickname
     })
 }
