+++ front/routes/chat/index.js
@@ -1,11 +1,12 @@
 const express = require('express')
 const chatController = require('./chatController.js')
+const { auth } = require('../../middleware/auth.js')
 const router = express.Router()
 
 
-router.use('/list', chatController.getChatList)
+router.use('/list', auth, chatController.getChatList)
 
-router.use('/room', chatController.getChatRoom)
+router.use('/room', auth, chatController.getChatRoom)
 
 
 module.exports = router
