+++ front/views/chat/chatRoom.html
@@ -14,6 +14,7 @@
     <div id="header">
         <h1 id="roomName">Room : {{roomName}}</h1>
         <h3 id="roomCount"></h3>
+        <input type="hidden" id="nickname" value="{{nickname}}">
     </div>
     <div id="firstMsg"></div>
     <div class="chat_wrap">
