+++ front/views/index.html
@@ -16,7 +16,7 @@
     <a href="/user/login"><button>로그인</button></a>
     <a href="/user/agree"><button>회원가입</button></a>
     <p></p>
-    <a href="javascript:openWindowPop('http://localhost:3001/chat/list', 'popup')" id="chat">채팅하기</a>
+    <a href="javascript:openWindowPop('/chat/list', 'popup')" id="chat">채팅하기</a>
 </body>
 
 </html>
